removed old files
separated db init
