+++ backend/matcha/repository.py
@@ -37,11 +37,10 @@ class UserRepository(Repository):
         )
 
     def update(self, user: User):
+        # TODO probably change condition from user_name to user_id
         self._engine.execute(
-            text('UPDATE Users SET first_name = :f, last_name = :l, gender = :g,'
-                 'preference = :p, biography = :b WHERE user_name = :u'),
-            u=user.get_user_name(), f=user.get_first_name(), l=user.get_last_name(), g=user.get_gender(),
-            p=user.get_preference(), b=user.get_biography()
+            text('UPDATE Users SET gender = :g,preference = :p, biography = :b WHERE user_name = :u'),
+            u=user.get_user_name(), g=user.get_gender(), p=user.get_preference(), b=user.get_biography()
         )
 
     def delete(self, user_name):
@@ -49,3 +48,28 @@ class UserRepository(Repository):
             text('DELETE FROM Users WHERE user_name = :u'),
             u=user_name
         )
+# def get_user_id(engine, user_name):
+#     result = engine.execute(
+#         text('SELECT user_id FROM Users WHERE user_name = :u'),
+#         u=user_name
+#     ).fetchone()
+#
+#     if result is not None:
+#         return result['user_id']
+#
+#
+# def get_user(user_name):
+#     engine = get_engine()
+#
+#     result = engine.execute(
+#         text('SELECT * FROM Users WHERE user_name = :u'),
+#         u=user_name
+#     ).fetchone()
+#
+#     return result
+#
+# def register_user(engine, user_name, password, first_name, last_name, email):
+#     engine.execute(
+#         text('INSERT INTO Users (user_name, password, first_name, last_name, email) VALUES (:u, :p, :f, :l, :e)'),
+#         u=user_name, p=generate_password_hash(password), f=first_name, l=last_name, e=email
+#     )
